Move out more inlay hints-related methods
